feat: migrate notion article detail page to app router
