+++ src/libs/constant.js
@@ -2,6 +2,7 @@ module.exports = {
   PAGE_TYPE_ARTICLE_SINGLE_PAGE: 'ArticleSinglePage',
   PAGE_TYPE_NOTION_SINGLE_PAGE: 'NotionSinglePage',
   PAGE_TYPE_NOTION_ARTICLE_LIST_PAGE: 'NotionArticleListPage',
+  PAGE_TYPE_NOTION_ARTICLE_DETAIL_PAGE: 'NotionArticleDetailPage',
   PAGE_TYPE_ERROR_PAGE: 'ErrorPage',
   END_TO_END_TEST_QUERY: 'e2e_test',
   FAILSAFE_PAGE_GENERATION_QUERY: 'fsg',
